+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [v1.5.0] - 2022-05-30
 
 - Upgrade to `protoc` 3.20.1 support.
 - Fix an issue where `buf` would fail if two or more roots contained
@@ -704,7 +704,7 @@ buf check breaking proto --against .git#branch=master,subdir=proto
 
 Initial beta release.
 
-[Unreleased]: https://github.com/bufbuild/buf/compare/v1.4.0...HEAD
+[v1.5.0]: https://github.com/bufbuild/buf/compare/v1.4.0...v1.5.0
 [v1.4.0]: https://github.com/bufbuild/buf/compare/v1.3.1...v1.4.0
 [v1.3.1]: https://github.com/bufbuild/buf/compare/v1.3.0...v1.3.1
 [v1.3.0]: https://github.com/bufbuild/buf/compare/v1.2.1...1.3.0